pdfimg2text/requirements.txt
# Extract images from PDF
PyMuPDF==1.21.1
# Wrapper for Tesseract OCR
pytesseract==0.3.10
# PIL
pillow==9.4.0
